Add respawn type check
Add a cvar, sm_auto_respawn_type, which allows for modifying the
behavior of the reset plugin. When set to 0 (default), if a repeat
AFK killer or spawn killer is detected, all player respawns will be
disabled. However, when set to 1, if a repeat AFK killer or spawn
killer is detected, only the player which is being repeatedly killed
will no longer be respawned. This cvar was added to allow more
flexibility especially when combined with the custom respawn
locations plugin which allows players to set their own custom respawn
location. Therefore, this cvar was added because it cannot be assumed
that all players will be killed by an AFK/spawn killer.

Other minor improvements:
- Optimize variable reset to occur only on client connection
- Refactor codebase