Fix admin respawn from not unblocking auto respawn
When players are killed by an AFK/spawn killer too fast, they are
subsequently blocked by the plugin from respawning again. This bug fix
resets the block status to allow for players to be auto-respawned
again after being force respawned by admins. This reset will occur
both via an admin saying !respawn <name> or if the admin respawns a
player by the admin menu. This change is based on the changed
assumption that once any AFK/spawn killer is detected, further auto-
respawns should be blocked. Rather, it is possible that a new respawn
location has become available and thus players should be allowed to
auto-respawn.